Create isPrime.py
This function will check to see if any number greater than 3 is prime or composite. Even numbers are automatically labelled composite.  Odd numbers, which are the only candidates for primes greater than 2 are subjected to a test where all odd numbers up to about N/3 are checked as factors.  This is because an odd number will only have odd factors and the biggest factor it may have is N/3. This function actually overshoots N/3 by one more number to be sure all relevant possible factors are checked.